REFACTOR ProductsListView component
