+++ styles/globals.css
@@ -30,109 +30,130 @@ a {
 }
 
 /* HEADER */
-.header {
+.header-container {
+  padding: 15px 10px;
+  border-bottom: 2px solid #f1f1f1;
+  display: flex;
+  justify-content: space-between;
 }
-  .header-container {
-    padding: 15px 10px;
-    border-bottom: 2px solid #f1f1f1;
+  .header-logo {
+    flex: 1;
+    height: 35px;
+    max-width: 180px;
+    margin-right: 20px;
+  }
+  .header-cart {
+    flex: 1;
+    background: #f1f1f1;
+    width: 80px;
+    margin-right: 20px;
+    border-radius: 20pt;
     display: flex;
     justify-content: space-between;
+    align-items: center;
+    padding-left: 10px;
   }
-    .header-logo {
-      flex: 1;
+    /* TODO: FIX when text in cart hav more than 5 numbers */
+    .header-cart span.count {
+      background: #F07439;
+      width: 35px;
       height: 35px;
-      max-width: 180px;
-      margin-right: 20px;
-    }
-    .header-cart {
-      flex: 1;
-      background: #f1f1f1;
-      width: 80px;
-      margin-right: 20px;
-      border-radius: 20pt;
+      border-radius: 50%;
       display: flex;
-      justify-content: space-between;
+      justify-content: center;
       align-items: center;
-      padding-left: 10px;
+      color: #fff;
+      font-size: 12px;
+      font-weight: bold;
     }
-      /* TODO: FIX when text in cart hav more than 5 numbers */
-      .header-cart span.count {
-        background: #F07439;
-        width: 35px;
-        height: 35px;
-        border-radius: 50%;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        color: #fff;
-        font-size: 12px;
-        font-weight: bold;
-      }
+
+/* FILTERS */
+.filter {
+  background: #f1f1f1;
+  padding: 10px 15px;
+  border-radius: 10pt 4pt 10pt 4pt;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 20px;
+}
+  .filter-title {
+    color: #afafaf;
+    font-size: 14px;
+  }
+  .filter-select select {
+    background: #f1f1f1;
+    border: none;
+    font-size: 14px;
+    color: #222222;
+    width: 100%;
+    
+  }
 
 /* PRODUCT CARD */
 .product {
-  margin-bottom: 20px;
-  display: flex;
-  flex-direction: row;
+margin-bottom: 20px;
+display: flex;
+flex-direction: row;
 }
-  .product-image {
+.product-image {
+  display: flex;
+  width: 160px;
+  height: 100;
+}
+  .product-image img {
+    border: 1px solid #f1f1f1;
+    border-radius: 10pt;
+    border-radius: 10pt 4pt 10pt 4pt;
+    width: 100%;
+    box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.1);
+  }
+.product-info {
+  flex: 1;
+  flex-direction: column;
+  padding: 0px 0px 0px 15px;
+}
+  .product-info-name {
+    margin: 0;
+    font-size: 20px;
+    font-weight: lighter;
+  }
+  .product-info-category {
+    margin: 0;
+    font-size: 14px;
+    font-weight: bold;
+    color: #afafaf;
+  }
+  .product-info-calification {
+    margin: 0;
+    font-size: 14px;
+    font-weight: lighter;
+  }
+  .product-info-footer {
     display: flex;
-    width: 160px;
-    height: 100;
+    flex-direction: row;
+    margin-top: 5px;
   }
-    .product-image img {
-      border: 1px solid #f1f1f1;
-      border-radius: 10pt;
-      border-radius: 10pt 4pt 10pt 4pt;
-      width: 100%;
-      box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.1);
-    }
-  .product-info {
-    flex: 1;
-    flex-direction: column;
-    padding: 0px 0px 0px 15px;
-  }
-    .product-info-name {
+    .product-info-price {
       margin: 0;
       font-size: 20px;
-      font-weight: lighter;
-    }
-    .product-info-category {
-      margin: 0;
-      font-size: 14px;
       font-weight: bold;
-      color: #afafaf;
+      color: #ffffff;
+      background: #F07439;
+      border-radius: 10pt 0 0 4pt;
+      padding: 5px 10px;
     }
-    .product-info-calification {
-      margin: 0;
-      font-size: 14px;
-      font-weight: lighter;
-    }
-    .product-info-footer {
+    .product-info-addtocart {
+      background: #f1f1f1;
+      color: #222222;
+      border: none;
+      border-radius: 0 4pt 10pt 0;
+      font-size: 12px;
+      font-weight: bold;
+      padding: 5px 10px;
       display: flex;
-      flex-direction: row;
-      margin-top: 5px;
+      align-items: center;
     }
-      .product-info-price {
-        margin: 0;
-        font-size: 20px;
-        font-weight: bold;
-        color: #ffffff;
-        background: #F07439;
-        border-radius: 10pt 0 0 4pt;
-        padding: 5px 10px;
-      }
-      .product-info-addtocart {
-        background: #f1f1f1;
-        color: #222222;
-        border: none;
-        border-radius: 0 4pt 10pt 0;
-        font-size: 12px;
-        font-weight: bold;
-        padding: 5px 10px;
-        display: flex;
-        align-items: center;
-      }
 
 .footer {
   background: #f1f1f1;
